m4-lobster: new "--modules" mode, yielding one lobster domain per selinux module
Each module has a port "ext" which is used for all connections into or out
of the module domain. Connections to an "ext" port have their final targets
specified with Lhs(subdom,port) or Rhs(subdom,port) annotations.

The implementation is as yet incomplete: Edges are not generated for
type_transition rules or for domtrans_pattern macros.